Make escape a second escape key and use dual shift for capslock.
I love having escape on the capslock key and thought moving capslock
to the escape key made sense as well. The problem is that many games
don't respect the X11's keymap, requiring me to use escape as escape
in games, which also caused capslock to be toggle. That was a mess.

I use CAPSLOCK SO RARELY that it hardly needs its own key, but I
noticed that xkb supports making both shift keys simultaneously turn
on capslock and then either shift key turns it off. Very slick.