Automatically (and recursively) expand ... to ../.. on tab or enter.
I've played with a few different versions of this trick since
encountering it in Oh-My-ZSH, but this one seems to be the most
practical.

I also modified it to only apply at the beginning of the line or after a
space, to keep it from breaking things like "git log A...B".