Add support for --from flag to git-branch-push
I often find myself having a mix of unpushed commits, some of which are
needed on the server right away while others aren't yet ready for the
limelight. In such cases, I do an interactive rebase to move the unready
commits to the end of the history and push just the stable commits to
the remote. Unfortunately, git-branch-push didn't support this as it
provided no way to specify which commit to push. Now it does with the
--from flag, which takes a mandatory commit ref.

This required a complete rewrite of the argument parsing code, which is
a big improvement.