git-branch-push: Whitelist flags that accept values
I'm not sure how I feel about this. I thought my original
implementation was really slick, but the problem was that it accepted
an equal sign after every flag. Usually this results in an error, but
if the value following the could be interpreted as a valid flag or
argument, it would be.

This new setup is a lot more correct and gives much better error
messages, but it's also a lot less clean.